+++ packages/python/tests/blueprints/multi-arch-buildx/fixtures/ci/m-image1.manifest.sh
@@ -6,5 +6,6 @@ docker buildx imagetools create \
   -t ghcr.io/repo-owner/m-image1:latest \
   -t ghcr.io/repo-owner/m-image1:v1 \
   -t ghcr.io/repo-owner/m-image1:v1.1 \
+  -t ghcr.io/repo-owner/m-image1:master \
   ghcr.io/repo-owner/amd64-m-image1:1.1.1 \
   ghcr.io/repo-owner/arm64-m-image1:1.1.1